+++ app/src/main/java/nars/game/sensor/VectorSensor.java
@@ -66,7 +66,7 @@ public abstract class VectorSensor extends AbstractSensor implements Iterable<Co
         }
         if (active > 0) {
             //post-commit phase
-            float aPri = pri.pri() / active;
+            float aPri = pri.pri() / (float) active;
             for (ComponentSignal s : this)
                 s.commit(aPri, wLoop);
         }
